Disable updating CI jobs temporarily
Temporarily disable updating CI jobs in ci_metrics_feed.py.

The current process is aimed at appending jobs, but not so efficient in the initial filling of the ducklake. Since we recently added all the the repositories in the duckdb org, we're now in the initial filling phase again.
To speed up this initial filling this will be done in a side process. (same logic, the rate limit is just distributed differently)